requirements.txt
tensorflow==2.15.0
numpy>=1.22
Pillow>=9.0
streamlit>=1.22
scikit-learn>=1.2
matplotlib>=3.5
